requirements.txt
streamlit
cohere
pillow
numpy 
requests 
tqdm
google-genai